+++ test/reuse-step-fn/playwright.config.ts
@@ -1,33 +1,36 @@
 import { defineConfig } from '@playwright/test';
 import { defineBddConfig } from 'playwright-bdd';
 
+const TAG = process.env.TAG; // @success, @error
+
 export default defineConfig({
   projects: [
     {
-      name: 'pw-style-success',
+      name: 'pw-style',
       testDir: defineBddConfig({
-        outputDir: '.features-gen/pw-style-success',
+        outputDir: `.features-gen/pw-style-${TAG}`,
         paths: ['features/*.feature'],
         require: ['steps-pw-style/*.ts'],
-        tags: '@success',
+        tags: TAG,
       }),
     },
     {
-      name: 'pw-style-invalid-invocation',
+      name: 'cucumber-style',
       testDir: defineBddConfig({
-        outputDir: '.features-gen/pw-style-invalid-invocation',
-        paths: ['features/*.feature'],
-        require: ['steps-pw-style/*.ts'],
-        tags: '@error',
-      }),
-    },
-    {
-      name: 'cucumber-style-success',
-      testDir: defineBddConfig({
-        outputDir: '.features-gen/cucumber-style-success',
+        outputDir: `.features-gen/cucumber-style-${TAG}`,
         paths: ['features/*.feature'],
         require: ['steps-cucumber-style/*.ts'],
-        tags: '@success',
+        tags: TAG,
+      }),
+    },
+    {
+      name: 'pw-style-world',
+      testDir: defineBddConfig({
+        outputDir: `.features-gen/pw-style-world-${TAG}`,
+        paths: ['features/*.feature'],
+        require: ['steps-pw-style-world/*.ts'],
+        verbose: true,
+        tags: TAG,
       }),
     },
   ],